Add files via upload
We'll be using a basic parameter implementation and tracking the line. First, we'll create a variable to hold your NFT and then create an empty array called "nftnum." Next, we'll use the "mint nft" function to create an NFT object using the parameter passed to it for its metadata. We'll also use the "boost" method and log a message to the console indicating that an NFT has been mounted. Then, we'll use the "least NFT" function to iterate over the NFT number using a for loop. This method will execute with some values assigned to it and print the total number of NFTs. Finally, we'll call the "mint NFT" function multiple times with different parameters to mint several NFTs, and then call the "enlist NFT" function to print the metadata of all minted NFTs. Overall, this code allows you to create NFT objects with metadata and track them easily.